Update DISCUSSION_READING_LIST.md
Add a video about filter bubbles to the "Equity, Identity, and Representation" topic. Filter bubbles are a new technologies for reinforcing the ideas that we already agree with. These echo chambers are a contributor to spreading prejudice and misinformation about other groups, and ultimately the maintenance of unequal power structures (or are they?).